requirements_cp311.txt
accelerate==1.10.1
diffusers==0.35.2
starlette<0.48 # gradio dose not support RFC9110 http status codes yet.
gradio==5.49.1
numpy==2.3.5
Pillow
psutil==7.1.0
PyYAML==6.0.3
Requests==2.32.5
torch==2.8.0
tqdm==4.67.1
transformers<4.57, >=4.54.0
nunchaku @ https://github.com/nunchaku-tech/nunchaku/releases/download/v1.0.2/nunchaku-1.0.2+torch2.8-cp311-cp311-linux_x86_64.whl
autoawq==0.2.9